+++ examples/llms/providers/customChatProvider.ts
@@ -1,6 +1,8 @@
 import {
   AsyncStream,
   BaseLLMTokenizeOutput,
+  EmbeddingOptions,
+  EmbeddingOutput,
   ExecutionOptions,
   GenerateOptions,
   LLMCache,
@@ -13,6 +15,7 @@ import { Emitter } from "bee-agent-framework/emitter/emitter";
 import { ChatLLM, ChatLLMGenerateEvents, ChatLLMOutput } from "bee-agent-framework/llms/chat";
 import { BaseMessage, Role } from "bee-agent-framework/llms/primitives/message";
 import { sum } from "remeda";
+import { NotImplementedError } from "bee-agent-framework/errors";
 
 export class CustomChatLLMOutput extends ChatLLMOutput {
   public readonly chunks: BaseMessage[] = [];
@@ -76,6 +79,10 @@ export class CustomChatLLM extends ChatLLM<CustomChatLLMOutput, CustomGenerateOp
   async meta(): Promise<LLMMeta> {
     // TODO: retrieve data about current model from the given provider API
     return { tokenLimit: Infinity };
+  }
+
+  async embed(input: BaseMessage[][], options?: EmbeddingOptions): Promise<EmbeddingOutput> {
+    throw new NotImplementedError();
   }
 
   async tokenize(input: BaseMessage[]): Promise<BaseLLMTokenizeOutput> {